A patch submitted by Chris Juettner in dev ML :
What is happening is the method is setting the dataResourceId in the map prior to the DataResource row being created. This is a problem because the dataResourceId doesn't exist yet. The result is that the uploaded file name is persisted with just the path and extension.  
 
For example:     
C:\eclipse\workspace\ofbiz\applications\content\uploads\1200348109149\.doc
 
So I added a call to the updateDataResource service following the createDataResource because the dataResourceId is now available. 


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/party@614945 13f79535-47bb-0310-9956-ffa450edef68