Fix a bug that seems to have been introduced at r934389 for (https://issues.apache.org/jira/browse/OFBIZ-3703) - OFBIZ-3703
This fix a bug in updatePartyContactMech. When you wanted to update only the extension of a phone number it did not get stored. 
I said "that seems to have been introduced" because it's working in stable demo.

I also remove a useless couple of lines at the beginning of the updatePartyContactMech service.

Finally I must say that I find very weird to have extension in PartyContactMech entity. It would be certainly more logical to have in TelecomNumber. Is it really used for anything else than phone number extension?

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/party@954583 13f79535-47bb-0310-9956-ffa450edef68