Removes tooltip="${uiLabelMap.CommonRequired}" from fields, https://issues.apache.org/jira/browse/OFBIZ-6691
Instead of a quiet S/R action, this was an unexpected pain in the ass. A considerable number of forms:
    EmployeeForms.xml
    EmploymentForms.xml
    EmplPositionForms.xml
    GlobalHRSettingForms.xml
    RecruitmentForms.xml
    QuoteForms.xml
    PartyForms.xml
    StoreForms.xml
    ContactForms.xml
    FixedAssetForms.xml
    TaskForms.xml
    scrumForms.xml
    TaskForms.xml
contained a large number of fields with 
  tooltip="${uiLabelMap.CommonRequired}"
but no 
  required-field="true"


Of course display and display-entity should not be concerned, I spotted a number of them in the lot.

I also removed a bunch of harcoded * at end of labels (those followed by required-field="true"* and spotted a number of others }*"  who still need to be carefully reviewed to see if required=true is not missing there :/

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/party@1709882 13f79535-47bb-0310-9956-ffa450edef68