Fixes a bug reported by Vinet at https://issues.apache.org/jira/browse/OFBIZ-6779 - "Wrong entity condition expression for fetching the view entity PartyAndContactMech."
In party module present in application component the file PartyContactMechServices.xml containing simple-method createPartyEmailAddress (meant for creating an email address for party) is using wrong entity condition expression for fetching the view entity PartyAndContactMech.
It is using from-field attribute for the comparison of a constant value instead of the value attribute.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/party@1721625 13f79535-47bb-0310-9956-ffa450edef68