Fixes: a bug introduced with one of (OFBIZ-8118)(OFBIZ-8122)(OFBIZ-8123)(OFBIZ-8124)(OFBIZ-8125)(OFBIZ-8126)(OFBIZ-8127)(OFBIZ-8128)(OFBIZ-8129)
I put the parenthesis here just because I copied from r1760059. In other cases I'll not except if it's *really* needed. I don't see why a simple space between Jira issues numbers would not be better, or not parenthesis at all for a sole issue. Is it for parsing and then why?

BTW, it's possible that there are more bugs like that (in screenlet, elsewere?) I did not check...

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/party@1760325 13f79535-47bb-0310-9956-ffa450edef68